requirements.txt
fastapi
uvicorn
openai>=1.35.0
httpx
requests
Pillow
google-genai
